requirements.txt
Django==4.2.7
django-crispy-forms==2.0
django-ratelimit==4.1.0
Pillow==10.0.1
python-magic==0.4.27
cryptography==41.0.4
bcrypt==4.0.1
psycopg2-binary==2.9.9 
django-environ==0.11.2
python-dotenv==1.0.0
django-debug-toolbar==4.2.0 
whitenoise==6.5.0  
gunicorn==21.2.0 
django-csp==3.7 
django-axes==6.1.1  